[REFACTOR] 공지사항 상세 관련 리팩토링
